# docs/gateway_swagger.yaml
basePath: /$gateway/api/
definitions:
  curd.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  curd.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  curd.ReplyData-types_Server:
    properties:
      data:
        $ref: '#/definitions/types.Server'
      error:
        type: string
    type: object
...
  curd.ReplyList-types_Server:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Server'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.DefaultDevice:
    properties:
      name:
        description: 名称
        type: string
      product_id:
        description: 产品ID
        type: string
      slave:
        description: 从站号
        type: integer
    type: object
...
  types.Server:
    properties:
      created:
        description: 创建时间
        type: string
      description:
        description: 说明
        type: string
      devices:
        items:
          $ref: '#/definitions/types.DefaultDevice'
        type: array
      disabled:
        type: boolean
      heartbeat:
        description: 心跳包
        type: string
      id:
        description: ID
        type: string
      name:
        description: 名称
        type: string
      poller_interval:
        description: 采集间隔
        type: integer
      poller_period:
        description: 采集周期
        type: integer
      port:
        description: 监听端口
        type: integer
      protocol_name:
        description: 协议 rtu tcp parallel-tcp
        type: string
      protocol_options:
        additionalProperties: {}
        description: 协议参数
        type: object
      running:
        type: boolean
      standalone:
        description: 单例模式（不支持注册）
        type: boolean
    type: object
info:
  contact: {}
  description: API文档
  title: 物联大师网关接口文档
  version: 1.0 版本
paths:
...
  /server/{id}:
    get:
      consumes:
      - application/json
      description: 获取服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 获取服务器
      tags:
      - server
    post:
      consumes:
      - application/json
      description: 修改服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 服务器信息
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/types.Server'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 修改服务器
      tags:
      - server
  /server/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 删除服务器
      tags:
      - server
  /server/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 禁用服务器
      tags:
      - server
  /server/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 启用服务器
      tags:
      - server
  /server/{id}/start:
    get:
      consumes:
      - application/json
      description: 启动服务端
      parameters:
      - description: 服务端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 启动服务端
      tags:
      - server
  /server/{id}/stop:
    get:
      consumes:
      - application/json
      description: 停止服务端
      parameters:
      - description: 服务端ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 停止服务端
      tags:
      - server
  /server/count:
    post:
      consumes:
      - application/json
      description: 查询服务器数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询服务器数量
      tags:
      - server
  /server/create:
    post:
      consumes:
      - application/json
      description: 创建服务器
      parameters:
      - description: 服务器信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Server'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Server'
      summary: 创建服务器
      tags:
      - server
...
  /server/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入服务端
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入服务端
      tags:
      - product
  /server/list:
    get:
      consumes:
      - application/json
      description: 查询服务器
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-types_Server'
      summary: 查询服务器
      tags:
      - server
  /server/search:
    post:
      consumes:
      - application/json
      description: 查询服务器
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-types_Server'
      summary: 查询服务器
      tags:
      - server
swagger: "2.0"